refactor: common modules
