+++ packages/topo-imagery-common/pyproject.toml
@@ -1,0 +1,19 @@
+[build-system]
+requires = ["uv_build>=0.11.14,<0.12.0"]
+build-backend = "uv_build"
+
+[project]
+name = "topo-imagery-common"
+version = "0.1.0"
+description = "Shared utilities for topo-imagery processing"
+requires-python = "~=3.12.3"
+dependencies = [
+    "boto3",
+    "linz-logger",
+    "py-multihash",
+    "python-dateutil",
+    "shapely",
+]
+
+[tool.uv]
+package = true
