+++ packages/topo-imagery-common/src/topo_imagery_common/files/tests/fs_s3_test.py
@@ -8,8 +8,8 @@ from mypy_boto3_s3 import S3Client
 from pytest import CaptureFixture, raises
 from pytest_subtests import SubTests
 
-from scripts.files.files_helper import ContentType
-from scripts.files.fs_s3 import exists, list_files_in_uri, read, write
+from topo_imagery_common.files.files_helper import ContentType
+from topo_imagery_common.files.fs_s3 import exists, list_files_in_uri, read, write
 
 
 @mock_aws
