+++ packages/topo-imagery-common/src/topo_imagery_common/files/tests/fs_test.py
@@ -10,7 +10,7 @@ from mypy_boto3_s3 import S3Client
 from pytest import CaptureFixture, raises
 from pytest_subtests import SubTests
 
-from scripts.files.fs import NoSuchFileError, read, write, write_all, write_sidecars
+from topo_imagery_common.files.fs import NoSuchFileError, read, write, write_all, write_sidecars
 
 
 def test_read_key_not_found_local() -> None:
