+++ pyproject.toml
@@ -20,6 +20,7 @@ dependencies = [
     "py-multihash",
     "shapely",
     "tifffile",
+    "topo-imagery-common",
 ]
 
 [dependency-groups]
@@ -45,6 +46,12 @@ line-length = 127
 [tool.uv]
 package = false
 default-groups = "all"
+
+[tool.uv.sources]
+topo-imagery-common = { workspace = true }
+
+[tool.uv.workspace]
+members = ["packages/*"]
 
 [tool.isort]
 line_length = 127
