+++ scripts/conftest.py
@@ -6,7 +6,7 @@ from typing import Iterator
 
 import pytest
 
-from scripts.datetimes import format_rfc_3339_datetime_string
+from topo_imagery_common.datetimes import format_rfc_3339_datetime_string
 from scripts.stac.imagery.collection_context import CollectionContext
