+++ scripts/fix_laz_header.py
@@ -8,10 +8,10 @@ from typing import Any, Iterable
 
 from linz_logger import get_log
 
-from scripts.cli.common_args import CommonArgumentParser
-from scripts.files.files_helper import ContentType
-from scripts.files.fs import copy, exists, read, write
-from scripts.logging.time_helper import time_in_ms
+from topo_imagery_common.cli.common_args import CommonArgumentParser
+from topo_imagery_common.files.files_helper import ContentType
+from topo_imagery_common.files.fs import copy, exists, read, write
+from topo_imagery_common.logging.time_helper import time_in_ms
 from scripts.pdal.pdal_commands import pdal_translate_add_proj_command, run_pdal
