+++ scripts/gdal/gdal_helper.py
@@ -8,11 +8,11 @@ from typing import cast
 
 from linz_logger import get_log
 
-from scripts.aws.aws_helper import is_s3
-from scripts.files.files_helper import get_file_name_from_path
-from scripts.files.fs import copy
+from topo_imagery_common.aws.aws_helper import is_s3
+from topo_imagery_common.files.files_helper import get_file_name_from_path
+from topo_imagery_common.files.fs import copy
 from scripts.gdal.gdalinfo import GdalInfo
-from scripts.logging.time_helper import time_in_ms
+from topo_imagery_common.logging.time_helper import time_in_ms
 
 
 class GDALExecutionException(Exception):
