+++ scripts/generate_hillshade.py
@@ -7,16 +7,16 @@ from multiprocessing import Pool
 
 from linz_logger import get_log
 
-from scripts.cli.cli_helper import InputParameterError, TileFiles, load_input_files
-from scripts.cli.common_args import CommonArgumentParser
-from scripts.datetimes import RFC_3339_DATETIME_FORMAT
-from scripts.files.files_helper import SUFFIX_JSON, ContentType, is_tiff
-from scripts.files.fs import exists, read, write, write_all
+from topo_imagery_common.cli.cli_helper import InputParameterError, TileFiles, load_input_files
+from topo_imagery_common.cli.common_args import CommonArgumentParser
+from topo_imagery_common.datetimes import RFC_3339_DATETIME_FORMAT
+from topo_imagery_common.files.files_helper import SUFFIX_JSON, ContentType, is_tiff
+from topo_imagery_common.files.fs import exists, read, write, write_all
 from scripts.gdal.gdal_commands import get_gdal_command, get_hillshade_command
 from scripts.gdal.gdal_helper import run_gdal
 from scripts.gdal.gdal_presets import CompressionPreset, HillshadePreset
 from scripts.json_codec import dict_to_json_bytes
-from scripts.logging.time_helper import time_in_ms
+from topo_imagery_common.logging.time_helper import time_in_ms
 from scripts.stac.imagery.create_stac import create_item
 from scripts.standardising import create_vrt
