+++ scripts/stac/imagery/tests/collection_test.py
@@ -16,9 +16,9 @@ from pytest import CaptureFixture, mark, param
 from pytest_subtests import SubTests
 
 from scripts.conftest import any_epoch_datetime_string, fake_linz_slug
-from scripts.files.files_helper import ContentType
-from scripts.files.fs import read
-from scripts.files.fs_s3 import write
+from topo_imagery_common.files.files_helper import ContentType
+from topo_imagery_common.files.fs import read
+from topo_imagery_common.files.fs_s3 import write
 from scripts.stac.imagery.collection import WARN_NO_PUBLISHED_CAPTURE_AREA, ImageryCollection, MissingMetadataError
 from scripts.stac.imagery.collection_context import CollectionContext
 from scripts.stac.imagery.item import ImageryItem, STACAsset
