+++ scripts/stac/imagery/tests/generators.py
@@ -3,7 +3,7 @@ from os import urandom
 from typing import Callable
 
 from scripts.stac.imagery.item import STACAsset, STACProcessing, STACProcessingSoftware
-from scripts.stac.util.checksum import multihash_as_hex
+from topo_imagery_common.files.checksum import multihash_as_hex
 
 
 def fixed_now_function(now: datetime) -> Callable[[], datetime]:
