+++ scripts/standardise_validate.py
@@ -5,7 +5,7 @@ from datetime import datetime, timezone
 
 from linz_logger import get_log
 
-from scripts.cli.cli_helper import (
+from topo_imagery_common.cli.cli_helper import (
     InputParameterError,
     load_input_files,
     str_to_bool,
@@ -14,10 +14,10 @@ from scripts.cli.cli_helper import (
     str_to_positive_int,
     valid_date,
 )
-from scripts.datetimes import RFC_3339_DATETIME_FORMAT, format_rfc_3339_nz_midnight_datetime_string
-from scripts.files.file_tiff import FileTiff
-from scripts.files.files_helper import SUFFIX_JSON, ContentType
-from scripts.files.fs import exists, write
+from topo_imagery_common.datetimes import RFC_3339_DATETIME_FORMAT, format_rfc_3339_nz_midnight_datetime_string
+from scripts.tiff.file_tiff import FileTiff
+from topo_imagery_common.files.files_helper import SUFFIX_JSON, ContentType
+from topo_imagery_common.files.fs import exists, write
 from scripts.gdal.gdal_helper import get_srs, get_vfs_path
 from scripts.json_codec import dict_to_json_bytes
 from scripts.stac.imagery.create_stac import create_item
