+++ scripts/tiff/tests/file_tiff_test.py
@@ -1,4 +1,4 @@
-from scripts.files.file_tiff import FileTiff, FileTiffErrorType
+from scripts.tiff.file_tiff import FileTiff, FileTiffErrorType
 from scripts.gdal.gdal_presets import CompressionPreset
 from scripts.gdal.tests.gdalinfo import add_band, add_palette_band, fake_gdal_info
